refactor(web): group submissions route components
